move typegen out of obsolete typescript/ dir
